requirements.txt
tensorflow==2.17.0
numpy==1.26.4
python-multipart==0.0.12
opencv-python==4.10.0.84
fastapi==0.115.0
uvicorn==0.32.0
torch==2.4.1
torchvision==0.19.1
pillow==10.4.0
pandas==2.2.3
scikit-learn==1.5.2
torch==2.8.0
